⚡ Optimize replaceTemplateRoles bulk insert via unnest
