--- backend/benchmark_mock.js
@@ -1,31 +1,0 @@
-async function runMockBenchmark() {
-  const numFields = 100;
-  const iterations = 10;
-
-  console.log(`Simulating benchmark with ${numFields} fields over ${iterations} iterations...`);
-
-  // N+1 approach
-  const startNPlus1 = performance.now();
-  for (let iter = 0; iter < iterations; iter++) {
-    for (let i = 0; i < numFields; i++) {
-      // Simulate network latency of query (e.g., 2ms per query)
-      await new Promise(resolve => setTimeout(resolve, 2));
-    }
-  }
-  const endNPlus1 = performance.now();
-  const avgNPlus1 = (endNPlus1 - startNPlus1) / iterations;
-  console.log(`Baseline (N+1) Average Time: ${avgNPlus1.toFixed(2)} ms`);
-
-  // Bulk Insert approach
-  const startBulk = performance.now();
-  for (let iter = 0; iter < iterations; iter++) {
-      // Simulate network latency of single bulk query (e.g., 5ms for large payload)
-      await new Promise(resolve => setTimeout(resolve, 5));
-  }
-  const endBulk = performance.now();
-  const avgBulk = (endBulk - startBulk) / iterations;
-  console.log(`Optimized (Bulk) Average Time: ${avgBulk.toFixed(2)} ms`);
-  console.log(`Improvement: ${((avgNPlus1 - avgBulk) / avgNPlus1 * 100).toFixed(2)}%`);
-}
-
-runMockBenchmark();
