+++ backend/package.json
@@ -38,7 +38,7 @@
     "node-cron": "^4.2.1",
     "pdf-lib": "^1.17.1",
     "pdfkit": "^0.17.2",
-    "pg": "^8.20.0",
+    "pg": "^8.16.0",
     "pg-hstore": "^2.3.4",
     "puppeteer": "^24.36.0",
     "resend": "^6.8.0",
