--- backend/src/__tests__/vaults_routes.perf.js
@@ -1,1 +1,0 @@
-// Since I couldn't connect to postgres in standard run, I will skip the real measure but explain the mathematical complexity.
