+++ backend/src/routes/recurring.routes.js
@@ -556,40 +556,30 @@ module.exports = (pool, authenticateJWT) => {
 
                 const invoiceId = invoiceResult.rows[0].id;
 
-                // Create invoice items using bulk insert
-                if (items && items.length > 0) {
-                    const values = [];
-                    const placeholders = [];
-                    let paramIndex = 1;
-
-                    for (let i = 0; i < items.length; i++) {
-                        const item = items[i];
-                        const itemTotal = (item.quantity || 1) * (item.unit_price || 0);
-                        const itemTax = itemTotal * ((item.tax_rate || 0) / 100);
-
-                        placeholders.push(`($${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++}, $${paramIndex++})`);
-
-                        values.push(
-                            invoiceId,
-                            req.organizationId,
-                            item.product_id || null,
-                            item.name,
-                            item.description || null,
-                            item.quantity || 1,
-                            item.unit_price || 0,
-                            item.tax_rate || 0,
-                            itemTax,
-                            itemTotal + itemTax,
-                            i
-                        );
-                    }
+                // Create invoice items
+                for (let i = 0; i < items.length; i++) {
+                    const item = items[i];
+                    const itemTotal = (item.quantity || 1) * (item.unit_price || 0);
+                    const itemTax = itemTotal * ((item.tax_rate || 0) / 100);
 
                     await client.query(`
                         INSERT INTO invoice_items (
                             invoice_id, organization_id, product_id, name, description,
                             quantity, unit_price, tax_rate, tax_amount, total, sort_order
-                        ) VALUES ${placeholders.join(', ')}
-                    `, values);
+                        ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
+                    `, [
+                        invoiceId,
+                        req.organizationId,
+                        item.product_id || null,
+                        item.name,
+                        item.description || null,
+                        item.quantity || 1,
+                        item.unit_price || 0,
+                        item.tax_rate || 0,
+                        itemTax,
+                        itemTotal + itemTax,
+                        i
+                    ]);
                 }
 
                 // Calculate next run date
